refactor: create bodies from parser functions go in separate file
